refactor: split app data and app hooks, fix tests
