+++ src/api/appData.ts
@@ -1,0 +1,98 @@
+import {
+  buildDeleteAppDataRoute,
+  buildDownloadFilesRoute,
+  buildGetAppDataRoute,
+  buildPatchAppDataRoute,
+  buildPostAppDataRoute,
+} from './routes';
+import { ApiData } from '../types';
+import { AppData, UUID } from '@graasp/sdk';
+import configureAxios from './axios';
+
+const axios = configureAxios();
+
+export const getAppData = async (args: ApiData): Promise<AppData[]> => {
+  const { token, itemId, apiHost } = args;
+  return axios
+    .get(`${apiHost}/${buildGetAppDataRoute(itemId)}`, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    })
+    .then(({ data }) => data);
+};
+
+export const postAppData = (
+  args: ApiData & {
+    body: unknown;
+  },
+): Promise<AppData> => {
+  const { token, itemId, apiHost, body } = args;
+  return axios
+    .post(`${apiHost}/${buildPostAppDataRoute({ itemId })}`, body, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    })
+    .then(({ data }) => data);
+};
+
+export const patchAppData = (args: ApiData & Partial<AppData> & { id: UUID }): Promise<AppData> => {
+  const { token, itemId, id, apiHost, data } = args;
+  return axios
+    .patch(
+      `${apiHost}/${buildPatchAppDataRoute({ itemId, id })}`,
+      { data },
+      {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      },
+    )
+    .then(({ data }) => data);
+};
+
+export const deleteAppData = (
+  args: ApiData & {
+    id: string;
+  },
+): Promise<AppData> => {
+  const { token, itemId, id, apiHost } = args;
+  return axios
+    .delete(`${apiHost}/${buildDeleteAppDataRoute({ itemId, id })}`, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    })
+    .then(({ data }) => data);
+};
+
+// todo: add return type of file
+// todo: add public route
+// because of the bearer token, it triggers an error on s3 on redirect because the request has two auth methods
+// https://github.com/axios/axios/issues/2855
+// https://stackoverflow.com/questions/50861144/reactjs-remove-http-header-before-redirect/51252434#51252434
+// so we removed automatic redirection for this endpoint
+export const getFileContent = async ({
+  id,
+  apiHost,
+  token,
+}: {
+  id: string;
+  apiHost: string;
+  token: string;
+}) => {
+  const url = await axios
+    .get(`${apiHost}/${buildDownloadFilesRoute(id)}`, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    })
+    .then(({ data }) => data);
+  return axios
+    .get(url, {
+      responseType: 'blob',
+      withCredentials: false,
+    })
+    .then(({ data }) => data);
+};
