+++ src/api/index.ts
@@ -1,2 +1,4 @@
 export * from './apps';
+export * from './appData';
+export * from './appAction';
 export * from './appSetting';
