+++ src/hooks/index.ts
@@ -1,6 +1,7 @@
 import { QueryClient } from '@tanstack/react-query';
 import { QueryClientConfig } from '../types';
-import configureAppsHooks from './appData';
+import configureAppsHooks from './app';
+import configureAppDataHooks from './appData';
 import configureAppSettingHooks from './appSetting';
 import configureAppActionHooks from './appAction';
 import configurePostMessageHooks from './postMessage';
@@ -8,6 +9,7 @@ import configurePostMessageHooks from './postMessage';
 export default (queryClient: QueryClient, queryConfig: QueryClientConfig) => {
   return {
     ...configureAppsHooks(queryClient, queryConfig),
+    ...configureAppDataHooks(queryClient, queryConfig),
     ...configurePostMessageHooks(queryClient, queryConfig),
     ...configureAppSettingHooks(queryClient, queryConfig),
     ...configureAppActionHooks(queryClient, queryConfig),
